+++ src/Exception/ExportPreviewValidationException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Omikron\Factfinder\Exception;
 
+use Throwable;
+
 class ExportPreviewValidationException extends \Exception
 {
     public function __construct($message = '', $code = 0, Throwable $previous = null)